refactor: split cipher with its helpers
