+++ ciphering-cli-tool/lib/cipher/cipher-helpers.js
@@ -1,0 +1,39 @@
+const {
+  CaesarEncoder,
+  CaesarDecoder,
+  Rot8Encoder,
+  Rot8Decoder,
+  AtbashTransformer,
+} = require('../transform')
+
+const { InputStream, OutputStream } = require('./cipher-io')
+
+const transformConsructorMap = {
+  'C1': CaesarEncoder,
+  'C0': CaesarDecoder,
+  'R1': Rot8Encoder,
+  'R0': Rot8Decoder,
+  'A': AtbashTransformer,
+}
+
+function getTransform(cmd) {
+  return new transformConsructorMap[cmd]()
+}
+
+function getTransformList(config) {
+  return config.split('-').map((cmd) => getTransform(cmd))
+}
+
+function getInputStream(input) {
+  return new InputStream(input)
+}
+
+function getOutputStream(input) {
+  return new OutputStream(input)
+}
+
+module.exports = {
+  getTransformList,
+  getInputStream,
+  getOutputStream,
+}
