+++ ciphering-cli-tool/lib/cipher/cipher.js
@@ -1,31 +1,11 @@
-const { pipeline } = require('stream/promises')
 const path = require('path')
-
-const { InputStream, OutputStream } = require('./cipher-io')
+const { pipeline } = require('stream/promises')
 
 const {
-  CaesarEncoder,
-  CaesarDecoder,
-  Rot8Encoder,
-  Rot8Decoder,
-  AtbashTransformer,
-} = require('../transform')
-
-const transformConsructorMap = {
-  'C1': CaesarEncoder,
-  'C0': CaesarDecoder,
-  'R1': Rot8Encoder,
-  'R0': Rot8Decoder,
-  'A': AtbashTransformer,
-}
-
-function getTransform(cmd) {
-  return new transformConsructorMap[cmd]()
-}
-
-function getTransformList(config) {
-  return config.split('-').map((cmd) => getTransform(cmd))
-}
+  getTransformList,
+  getInputStream,
+  getOutputStream,
+} = require('./cipher-helpers')
 
 async function cipher({
   config,
@@ -37,9 +17,9 @@ async function cipher({
   const outputPath = (output) ? path.resolve(output) : null
 
   await pipeline(
-    new InputStream(inputPath),
+    getInputStream(inputPath),
     ...getTransformList(config),
-    new OutputStream(outputPath)
+    getOutputStream(outputPath)
   )
 }
